EditPage moved to pages package
